Summary of this project
SmartShop is a fictional mid-sized retail company that wants to expand its business by launching an online store. As they anticipate an increase in web traffic due to marketing campaigns and seasonal sales, SmartShop requires a robust and reliable web infrastructure to ensure smooth customer experiences. With minimal in-house IT staff and a limited
budget, they seek a cost-effective solution that can be manually adjusted for growth as needed, providing high availability and essential security while keeping operational complexity low.

As a cloud engineer, I am committed to bringing this into life by building a scalable web infrastructure tailored to meet this need.

THE AIMS OF THIS PROJECT INCLUDE CREATING A WEB INFRASTRUCTURE THAT:
. Support anticipated growth with flexible resource allocation.
• Ensure high availability and consistent user experience.
• Implement strong security measures to protect customer data.
• Be cost-effective and optimized for performance.